intemrid

+++ src/main/scala/windward/simulation/physical/weather/WeatherGenerator.scala
@@ -17,13 +17,13 @@ object WeatherGenerator {
      * @param windSpeed
      * @return
      */
-    def initUniformWind(width: SimUnit, height: SimUnit, windDirection: Int, windSpeed: Int): WeatherMap = {
+    def initUniformWind(width: SimUnit, height: SimUnit, windDirection: Int, windSpeed: Float): WeatherMap = {
 
-        val speed: Array[Array[Int]] = new Array[Array[Int]](width.toCellUnit.toInt);
+        val speed: Array[Array[Float]] = new Array[Array[Float]](width.toCellUnit.toInt);
         val direction: Array[Array[Int]] = new Array[Array[Int]](width.toCellUnit.toInt);
 
         for (rowIndex <- 0 until width.toInt) {
-            speed(rowIndex) = new Array[Int](height.toCellUnit.toInt);
+            speed(rowIndex) = new Array[Float](height.toCellUnit.toInt);
             direction(rowIndex) = new Array[Int](height.toCellUnit.toInt);
 
             for (colIndex <- 0 until height.toCellUnit.toInt) {
@@ -43,17 +43,17 @@ object WeatherGenerator {
      * @param maxSpeed
      * @return
      */
-    def initRandomWind(width: SimUnit, height: SimUnit, minSpeed: Int, maxSpeed: Int): WeatherMap = {
+    def initRandomWind(width: SimUnit, height: SimUnit, minSpeed: Float, maxSpeed: Float): WeatherMap = {
         val rnd: Random = new Random();
-        val speed: Array[Array[Int]] = new Array[Array[Int]](width.toCellUnit.toInt);
+        val speed: Array[Array[Float]] = new Array[Array[Float]](width.toCellUnit.toInt);
         val direction: Array[Array[Int]] = new Array[Array[Int]](width.toCellUnit.toInt);
 
         for (rowIndex <- 0 until width.toCellUnit.toInt) {
-            speed(rowIndex) = new Array[Int](height.toCellUnit.toInt);
+            speed(rowIndex) = new Array[Float](height.toCellUnit.toInt);
             direction(rowIndex) = new Array[Int](height.toCellUnit.toInt);
 
             for (colIndex <- 0 until height.toCellUnit.toInt) {
-                speed(rowIndex)(colIndex) = rnd.nextInt(maxSpeed - minSpeed) + minSpeed;;
+                speed(rowIndex)(colIndex) = rnd.nextInt((maxSpeed - minSpeed).toInt) + minSpeed;;
                 direction(rowIndex)(colIndex) = rnd.nextInt(360);;
             }
         }
@@ -69,25 +69,25 @@ object WeatherGenerator {
      * @param startingSpeed
      * @return
      */
-    def initGradientWind(width: SimUnit, height: SimUnit, windDirection: Int, startingSpeed: Int): WeatherMap = {
+    def initGradientWind(width: SimUnit, height: SimUnit, windDirection: Int, startingSpeed: Float): WeatherMap = {
 
         def distance(x: Int, y: Int) = {
             Math.sqrt(Math.pow(0 - x, 2) + Math.pow(0 - y, 2));
         }
 
         val furthestDistance = distance(width.toCellUnit.toInt, height.toCellUnit.toInt);
-        val speed: Array[Array[Int]] = new Array[Array[Int]](width.toCellUnit.toInt);
+        val speed: Array[Array[Float]] = new Array[Array[Float]](width.toCellUnit.toInt);
         val direction: Array[Array[Int]] = new Array[Array[Int]](width.toCellUnit.toInt);
 
         for (rowIndex <- 0 until width.toCellUnit.toInt) {
-            speed(rowIndex) = new Array[Int](height.toCellUnit.toInt);
+            speed(rowIndex) = new Array[Float](height.toCellUnit.toInt);
             direction(rowIndex) = new Array[Int](height.toCellUnit.toInt);
 
             for (colIndex <- 0 until height.toCellUnit.toInt) {
                 val strength = (Math.max(1, distance(rowIndex, colIndex)) / furthestDistance);
 
                 direction(rowIndex)(colIndex) = windDirection;
-                speed(rowIndex)(colIndex) = Math.max(startingSpeed - (strength * startingSpeed).toInt, 0);
+                speed(rowIndex)(colIndex) = Math.max(startingSpeed - (strength * startingSpeed), 0).toFloat;
             }
         }
         new WeatherMap(speed, direction);
