+++ src/main/scala/windward/simulation/physical/weather/WeatherMap.scala
@@ -5,13 +5,13 @@ import windward.simulation.units.CellUnit
 /**
  * @author Janoky Laszlo Viktor <janoky.laszlo@bmeautsoft.hu>
  */
-class WeatherMap(var speed: Array[Array[Int]], var direction: Array[Array[Int]]) {
+class WeatherMap(var speed: Array[Array[Float]], var direction: Array[Array[Int]]) {
 
     def getDirection(x: CellUnit, y : CellUnit) : Int = {
         direction(x.value)(y.value)
     }
 
-    def getSpeed(x: CellUnit, y : CellUnit) : Int = {
+    def getSpeed(x: CellUnit, y : CellUnit) : Float = {
         speed(x.value)(y.value)
     }
 }
